requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --constraint=/home/joakim/work/odin-api/odin-infrastructure/requirements-dev.txt --no-annotate --output-file=requirements.txt requirements.in
#
attrs==25.3.0
aws-cdk-asset-awscli-v1==2.2.237
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-cloud-assembly-schema==44.2.0
aws-cdk-lib==2.200.1
cattrs==24.1.3
constructs==10.4.2
exceptiongroup==1.3.0
importlib-resources==6.5.2
jsii==1.112.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==2.13.3
typing-extensions==4.14.0
